+++ skyportal/tests/flaky/test_frontend_flaky.py
@@ -1,3 +1,4 @@
+import datetime
 import os
 import uuid
 import pandas as pd
@@ -872,3 +873,77 @@ def test_gcn_request(driver, user, super_admin_token, public_group):
         f'//a[contains(@data-testid, "observationGcn_{instrument_id}")]',
         scroll_parent=True,
     )
+
+
+@pytest.mark.flaky(reruns=2)
+def test_candidate_date_filtering(
+    driver,
+    user,
+    public_candidate,
+    public_filter,
+    public_group,
+    upload_data_token,
+    ztf_camera,
+):
+    candidate_id = str(uuid.uuid4())
+    for i in range(5):
+        status, data = api(
+            "POST",
+            "candidates",
+            data={
+                "id": f"{candidate_id}_{i}",
+                "ra": 234.22,
+                "dec": -22.33,
+                "redshift": 3,
+                "altdata": {"simbad": {"class": "RRLyr"}},
+                "transient": False,
+                "ra_dis": 2.3,
+                "filter_ids": [public_filter.id],
+                "passed_at": str(datetime.datetime.utcnow()),
+            },
+            token=upload_data_token,
+        )
+        assert status == 200
+
+        status, data = api(
+            "POST",
+            "photometry",
+            data={
+                "obj_id": f"{candidate_id}_{i}",
+                "mjd": 58000.0,
+                "instrument_id": ztf_camera.id,
+                "flux": 12.24,
+                "fluxerr": 0.031,
+                "zp": 25.0,
+                "magsys": "ab",
+                "filter": "ztfr",
+                "group_ids": [public_group.id],
+            },
+            token=upload_data_token,
+        )
+        assert status == 200
+
+    driver.get(f"/become_user/{user.id}")
+    driver.get("/candidates")
+    driver.click_xpath(
+        f'//*[@data-testid="filteringFormGroupCheckbox-{public_group.id}"]',
+        wait_clickable=False,
+    )
+    start_date_input = driver.wait_for_xpath("//input[@id='startDatePicker']")
+    start_date_input.clear()
+    start_date_input.send_keys("12/01/2020 12:00 p")
+    end_date_input = driver.wait_for_xpath("//input[@id='endDatePicker']")
+    end_date_input.clear()
+    end_date_input.send_keys("12/01/2020 12:00 p")
+    submit_button = driver.wait_for_xpath_to_be_clickable('//button[text()="Search"]')
+    driver.scroll_to_element_and_click(submit_button)
+    for i in range(5):
+        driver.wait_for_xpath_to_disappear(
+            f'//a[@data-testid="{candidate_id}_{i}"]', 10
+        )
+    end_date_input.clear()
+    end_date_input.send_keys("12/01/2090 12:00 p")
+    submit_button = driver.wait_for_xpath_to_be_clickable('//button[text()="Search"]')
+    driver.scroll_to_element_and_click(submit_button)
+    for i in range(5):
+        driver.wait_for_xpath(f'//a[@data-testid="{candidate_id}_{i}"]', 10)
